requirements.txt
Pillow>=9.0.0
pyinstaller>=5.0.0  # 仅编译时需要
auto-py-to-exe>=2.0.0  # 仅使用图形界面编译时需要